Clean up numbers (reduce casts)

+++ Holo/Media/Providers/ISourceProvider.cs
@@ -98,8 +98,8 @@ public interface ISourceProvider
     unsafe Bitmap* GetVisualization(
         int width,
         int height,
-        float pixelsPerMs,
-        float amplitudeScale,
+        double pixelsPerMs,
+        double amplitudeScale,
         long startTime,
         long frameTime
     );
