+++ Holo/Media/Providers/MizukiSourceProvider.cs
@@ -147,8 +147,8 @@ public unsafe class MizukiSourceProvider : ISourceProvider
     public Bitmap* GetVisualization(
         int width,
         int height,
-        float pixelsPerMs,
-        float amplitudeScale,
+        double pixelsPerMs,
+        double amplitudeScale,
         long startTime,
         long frameTime
     )
@@ -324,8 +324,8 @@ internal static unsafe partial class External
         GlobalContext* context,
         int width,
         int height,
-        float pixelsPerMs,
-        float amplitudeScale,
+        double pixelsPerMs,
+        double amplitudeScale,
         long startTime,
         long frameTime
     );
